makes a views file for the cog
